:recycle: Refactor the project
* changed the project structure
* replaced ava and sinon with jest
* improved liniting
