+++ .editorconfig
@@ -4,5 +4,5 @@ root = true
 charset = utf-8
 end_of_line = lf
 insert_final_newline = true
-indent_size = 2
+indent_size = 4
 indent_style = space
