+++ .gitignore
@@ -1,3 +1,7 @@
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
 build/
 node_modules/
 settings.yaml
