+++ .travis.yml
@@ -6,8 +6,13 @@ node_js:
   - "8"
 
 cache:
+  yarn: true
   directories:
     - node_modules
 
 install:
-  - npm install
+  - yarn install
+
+script:
+  - yarn lint
+  - yarn test
