+++ src/bootstrap.ts
@@ -1,5 +1,5 @@
-import {validateLectureIds} from './validators';
-import {generateLecturesUserStyles} from './generators';
+import { validateLectureIds } from './validator/lecture-ids';
+import { generateLecturesUserStyles } from './generator/user-styles';
 
 validateLectureIds();
 generateLecturesUserStyles();
