+++ src/generator/user-styles.ts
@@ -1,0 +1,23 @@
+import * as fse from 'fs-extra';
+import * as rif from 'replace-in-file';
+import { templateFilePath, buildFilePath } from '../config';
+import { getLecturesUrlPrefixes } from '../helper/url-prefixes';
+
+export function generateLecturesUserStyles(): void {
+    fse.copy(templateFilePath, buildFilePath, (error: Error | null) => {
+        if (error) {
+            throw new Error('Could not copy the source file into the given destination');
+        }
+
+        const options = {
+            files: buildFilePath,
+            from: /URL_PREFIXES/g,
+            to: getLecturesUrlPrefixes(),
+            encoding: 'utf8',
+        };
+
+        rif(options)
+            .then(() => console.log('Build finished successfully!'))
+            .catch(console.error);
+    });
+}
