+++ src/interface/settings.ts
@@ -1,0 +1,3 @@
+export interface Settings {
+    lectureIds: number[];
+}
