+++ tslint.json
@@ -1,0 +1,50 @@
+{
+    "extends": "tslint:recommended",
+    "rules": {
+        "array-type": [
+            true,
+            "array"
+        ],
+        "interface-name": [
+            true,
+            "never-prefix"
+        ],
+        "max-classes-per-file": [
+            false
+        ],
+        "max-line-length": [
+            true,
+            200
+        ],
+        "no-bitwise": [
+            false
+        ],
+        "no-console": [
+            false,
+            "log"
+        ],
+        "no-empty": false,
+        "no-namespace": [
+            true,
+            "allow-declarations"
+        ],
+        "no-unused-expression": [
+            true,
+            "allow-fast-null-checks"
+        ],
+        "no-var-keyword": false,
+        "object-literal-shorthand": false,
+        "object-literal-sort-keys": false,
+        "only-arrow-functions": false,
+        "ordered-imports": false,
+        "quotemark": [
+            true,
+            "single",
+            "avoid-escape"
+        ],
+        "variable-name": [
+            true,
+            "ban-keywords"
+        ]
+    }
+}
